Backend/requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
python-dotenv
openai
chromadb
langsmith
flask
flask-cors
gunicorn
faiss-cpu 